complete classapply,lesson, into class

+++ eGov-School/src/main/java/com/school/dto/ClassApplyVO.java
@@ -34,5 +34,7 @@ public class ClassApplyVO {
 	
 	// ClassApplyVO.java에 추가
 	private String claName; // JSP ${apply.claName} 매핑용
-	private int progress;   // JSP ${apply.progress} 매핑용
+	private double progress;   // JSP ${apply.progress} 매핑용
+	
+	
 }
